+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - downloaded and decompressed `gz`-files are now correctly renamed
 - you can now select whether you want a prefix for downloaded log files or 
   not for each log file source in the settings
+- log files are now also getting downloaded when double clicking on them
 
 ## 16.04.2
 - **switched LogAnalyzer release versioning** to a more *rolling release